refactor: tighten CSP, consolidate modals, extract validation/stats helpers
Multi-agent review findings rolled into a single PR. Three coordinated changes:

Security
- public/_headers: restrict connect-src from `https:` to
  `https://api.anthropic.com` so an XSS payload can't exfiltrate to
  arbitrary HTTPS endpoints.

UI consolidation
- src/ui/modalShell.ts: add `onClose` hook so Promise-returning modals
  can resolve on Escape / click-outside / ✕.
- src/ui/exportOptionsDialog.ts, src/ui/importPreview.ts: rewrite on
  top of modalShell instead of hand-rolling overlay + Escape handler +
  cleanup. ~150 LOC removed, single source of truth for modal chrome.
- src/ui/styleConstants.ts: extract button + overlay class strings
  and z-index values.
- src/ui/htmlUtils.ts: shared escapeHtml() so callers don't reinvent
  (or skip) escaping.

Main.ts decomposition (5489 → 5138 LOC)
- src/validation/apiValidation.ts: pure runtime type assertions
  (assertString, assertNumber, guard, etc.) used across the
  window.partwright API.
- src/geometry/statsComputation.ts: computeGeometryStats,
  computeStatDiff, checkAssertions, GeometryAssertions, simpleHash.
- src/main.ts: add file-level docstring explaining what main owns
  and where the extracted helpers live. Comment why the drag/drop
  document listeners aren't cleaned up.
- src/ai/tools.ts: add tool-flow docstring.

Verified: full Playwright suite (34 tests) passes; production build
succeeds with no new warnings.
