+++ src/geometry/statsComputation.ts
@@ -1,0 +1,218 @@
+// Pure helpers that compute the machine-readable geometry-data payload
+// surfaced via window.partwright. Separated from main.ts so the math /
+// shape / fields can be tested and edited without touching the editor
+// glue. All functions take their inputs as arguments — no closures over
+// editor state.
+
+import { sliceAtZ, getBoundingBox } from './crossSection';
+import { getUnits } from './units';
+import type { MeshData } from './types';
+
+/** Manifold runtime type — the manifold-3d package does not export
+ *  precise TS types for the WASM-backed object, so we treat it as an
+ *  opaque value with method probes. */
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+type Manifold = any;
+
+/** Stable, short content hash. Used to stamp the geometry-data payload so
+ *  callers can detect when an unchanged code re-render produced the same
+ *  output. Not cryptographic — collisions are acceptable. */
+export function simpleHash(str: string): string {
+  let h = 0;
+  for (let i = 0; i < str.length; i++) {
+    h = ((h << 5) - h + str.charCodeAt(i)) | 0;
+  }
+  return (h >>> 0).toString(16).padStart(8, '0');
+}
+
+export function computeGeometryStats(
+  manifold: Manifold,
+  meshData: MeshData,
+  executionTimeMs?: number,
+  sourceCode?: string,
+): Record<string, unknown> {
+  const bbox = getBoundingBox(manifold);
+
+  let volume = 0;
+  let surfaceArea = 0;
+  try {
+    volume = manifold.volume();
+    surfaceArea = manifold.surfaceArea();
+  } catch {
+    // fallback if methods unavailable
+  }
+
+  const centroid = bbox
+    ? [(bbox.min[0] + bbox.max[0]) / 2, (bbox.min[1] + bbox.max[1]) / 2, (bbox.min[2] + bbox.max[2]) / 2]
+    : null;
+
+  const dimensions = bbox
+    ? [bbox.max[0] - bbox.min[0], bbox.max[1] - bbox.min[1], bbox.max[2] - bbox.min[2]]
+    : null;
+
+  let componentCount = 1;
+  try {
+    const parts = manifold.decompose();
+    componentCount = parts.length;
+    for (const p of parts) p.delete();
+  } catch {
+    // fallback
+  }
+
+  let isManifold = true;
+  let manifoldStatus: string | null = null;
+  try {
+    const s = manifold.status();
+    isManifold = s === 0 || s === 'NoError';
+    if (!isManifold) {
+      manifoldStatus = String(s);
+    }
+  } catch {
+    // fallback
+  }
+
+  const quartileSlices: Record<string, { z: number; area: number; contours: number }> = {};
+  if (bbox) {
+    const zRange = bbox.max[2] - bbox.min[2];
+    for (const pct of [25, 50, 75]) {
+      const z = bbox.min[2] + zRange * (pct / 100);
+      const s = sliceAtZ(manifold, z);
+      if (s) {
+        quartileSlices[`z${pct}`] = { z, area: s.area, contours: s.polygons.length };
+      }
+    }
+  }
+
+  return {
+    status: 'ok' as const,
+    vertexCount: meshData.numVert,
+    triangleCount: meshData.numTri,
+    boundingBox: bbox ? {
+      x: [bbox.min[0], bbox.max[0]],
+      y: [bbox.min[1], bbox.max[1]],
+      z: [bbox.min[2], bbox.max[2]],
+      dimensions,
+    } : null,
+    centroid,
+    volume,
+    surfaceArea,
+    genus: (() => { try { return manifold.genus(); } catch { return null; } })(),
+    isManifold,
+    ...(manifoldStatus ? { manifoldStatus } : {}),
+    componentCount,
+    crossSections: quartileSlices,
+    unit: getUnits(),
+    executionTimeMs: executionTimeMs ?? null,
+    codeHash: sourceCode ? simpleHash(sourceCode) : null,
+  };
+}
+
+export function computeStatDiff(prev: Record<string, unknown>, next: Record<string, unknown>): Record<string, unknown> {
+  const diff: Record<string, unknown> = {};
+
+  const numericFields = ['volume', 'surfaceArea', 'vertexCount', 'triangleCount', 'genus', 'componentCount'];
+  for (const field of numericFields) {
+    const from = prev[field] as number;
+    const to = next[field] as number;
+    if (from !== undefined && to !== undefined) {
+      const delta = to - from;
+      if (delta === 0) {
+        diff[field] = { from, to, delta: 'unchanged' };
+      } else {
+        const pct = from !== 0 ? ((delta / from) * 100).toFixed(1) : null;
+        diff[field] = {
+          from, to,
+          delta: `${delta > 0 ? '+' : ''}${Math.round(delta)}${pct ? ` (${delta > 0 ? '+' : ''}${pct}%)` : ''}`,
+        };
+      }
+    }
+  }
+
+  const prevBB = prev.boundingBox as Record<string, unknown> | null;
+  const nextBB = next.boundingBox as Record<string, unknown> | null;
+  if (prevBB?.dimensions && nextBB?.dimensions) {
+    diff.boundingBox = { dimensions: { from: prevBB.dimensions, to: nextBB.dimensions } };
+  }
+
+  return diff;
+}
+
+export interface GeometryAssertions {
+  minVolume?: number;
+  maxVolume?: number;
+  isManifold?: boolean;
+  maxComponents?: number;
+  genus?: number;
+  minGenus?: number;
+  maxGenus?: number;
+  minBounds?: [number, number, number];
+  maxBounds?: [number, number, number];
+  minTriangles?: number;
+  maxTriangles?: number;
+  /** Proportion range assertions: { widthToDepth: [min, max], widthToHeight: [min, max], depthToHeight: [min, max] } */
+  boundsRatio?: {
+    widthToDepth?: [number, number];
+    widthToHeight?: [number, number];
+    depthToHeight?: [number, number];
+  };
+  /** Optional notes to attach to this version (design rationale, user feedback, etc.) */
+  notes?: string;
+}
+
+export function checkAssertions(stats: Record<string, unknown>, assertions: GeometryAssertions): string[] {
+  const failures: string[] = [];
+  const v = stats.volume as number;
+  const tc = stats.triangleCount as number;
+  const cc = stats.componentCount as number;
+  const g = stats.genus as number | null;
+  const im = stats.isManifold as boolean;
+  const bb = stats.boundingBox as { dimensions?: number[] } | null;
+
+  if (assertions.minVolume !== undefined && v < assertions.minVolume)
+    failures.push(`volume ${v.toFixed(1)} < minVolume ${assertions.minVolume}`);
+  if (assertions.maxVolume !== undefined && v > assertions.maxVolume)
+    failures.push(`volume ${v.toFixed(1)} > maxVolume ${assertions.maxVolume}`);
+  if (assertions.isManifold !== undefined && im !== assertions.isManifold)
+    failures.push(`isManifold is ${im}, expected ${assertions.isManifold}`);
+  if (assertions.maxComponents !== undefined && cc > assertions.maxComponents)
+    failures.push(`componentCount ${cc} > maxComponents ${assertions.maxComponents}`);
+  if (assertions.genus !== undefined && g !== assertions.genus)
+    failures.push(`genus ${g} !== expected ${assertions.genus}`);
+  if (assertions.minGenus !== undefined && (g === null || g < assertions.minGenus))
+    failures.push(`genus ${g} < minGenus ${assertions.minGenus}`);
+  if (assertions.maxGenus !== undefined && (g === null || g > assertions.maxGenus))
+    failures.push(`genus ${g} > maxGenus ${assertions.maxGenus}`);
+  if (assertions.minTriangles !== undefined && tc < assertions.minTriangles)
+    failures.push(`triangleCount ${tc} < minTriangles ${assertions.minTriangles}`);
+  if (assertions.maxTriangles !== undefined && tc > assertions.maxTriangles)
+    failures.push(`triangleCount ${tc} > maxTriangles ${assertions.maxTriangles}`);
+  if (assertions.minBounds && bb?.dimensions) {
+    const d = bb.dimensions;
+    for (let i = 0; i < 3; i++) {
+      if (d[i] < assertions.minBounds[i])
+        failures.push(`dimension ${['X', 'Y', 'Z'][i]} ${d[i].toFixed(1)} < minBounds ${assertions.minBounds[i]}`);
+    }
+  }
+  if (assertions.maxBounds && bb?.dimensions) {
+    const d = bb.dimensions;
+    for (let i = 0; i < 3; i++) {
+      if (d[i] > assertions.maxBounds[i])
+        failures.push(`dimension ${['X', 'Y', 'Z'][i]} ${d[i].toFixed(1)} > maxBounds ${assertions.maxBounds[i]}`);
+    }
+  }
+  if (assertions.boundsRatio && bb?.dimensions) {
+    const [w, dep, h] = bb.dimensions;
+    const ratios: { name: string; value: number; range?: [number, number] }[] = [
+      { name: 'widthToDepth', value: w / dep, range: assertions.boundsRatio.widthToDepth },
+      { name: 'widthToHeight', value: w / h, range: assertions.boundsRatio.widthToHeight },
+      { name: 'depthToHeight', value: dep / h, range: assertions.boundsRatio.depthToHeight },
+    ];
+    for (const r of ratios) {
+      if (r.range) {
+        if (r.value < r.range[0]) failures.push(`${r.name} ratio ${r.value.toFixed(2)} < min ${r.range[0]}`);
+        if (r.value > r.range[1]) failures.push(`${r.name} ratio ${r.value.toFixed(2)} > max ${r.range[1]}`);
+      }
+    }
+  }
+  return failures;
+}
